Update `Event` to set a custom payload type
This updates event to set a custom payload type when it builds a
job. This is part of supporting events and avoiding routing issues
with jobs. By always settings its payload type to something
different its not possible for a job and event job to have the
same route name.

This also changes the `Event.build` method to take an options hash
on the end instead of its published at timestamp. This is to give
it a consistent feel as `Job` which now takes an options hash. This
doesn't change any of the functionality, just the interface.