Add `Qs.publish`
This adds the ability to publish events using Qs. This creates a
dispatch job and adds it to the configured dispatcher queue. This
will allow the dispatcher to find subscriptions for the event and
add event jobs to each of the subscribed queues.

This also adds `published_events` to Qs. This is a test helper for
checking what events have been published. The `TestClient` will
add dispatch jobs to the dispatcher queues enqueued jobs which
allows asking for event the dispatch job represents.

This also includes a few test cleanups. Job factories now allow
passing params so their values can be set to a specific value if
needed. This also updates the base `Client` enqueue tests to check
that it uses the queue that it is passed. This was accidentally
missed when it was originally developed.