Add `event` to `Queue`
This adds an `event` method to `Queue` for configuring what events
a queue listens for and can handle. This works similarly to the
`job` method, the only difference is it builds the job name using
the event channel and event name. This will allow an event job to
be routed to the configured handler correctly.

This also adds an `event_handler_ns` which functions the same as
the `job_handler_ns`. Event handler class names will be prefixed
with it if set.

Finally, this also includes an additional test to check that the
job handler ns is ignored for a top-level handler class name
(prefixed with `::`).