Split `TestHelpers`, one for `JobHandler` and one for `EventHandler`
This splits the test helpers into two mixins, one for `JobHandler`
and one for `EventHandler`. This is so each can provide a different
`test_runner` method and in general, different test helpers as
needed.

This renames the `TestRunner` to `JobTestRunner` and adds an
`EventTestRunner` that inherits from it. The `EventTestRunner`
checks that its passed an `EventHandler` and also sets up passing
an event job. `EventHandler`, unlike `JobHandler`, requires its
runner to have an event job. To handle this, the `EventTestRunner`
will build an event job and super it to the `TestRunner`. It also
handles params so it can easily be passed event params directly.

This also fixes a minor issue with how the `Runner` set its
attributes and built its handler. Since event handlers require
their runner to have an event job when they are built, the base
`Runner` needs to set its attributes before it builds an instance
of the handler.

All of this is part of adding an events system to Qs.