Store queues as event subscribers in redis
This updates queues to store themselves as subscribers for their
events in redis. This is part of the events system. This will be
used by the dispatcher when it receives a dispatch job from an
event publish. The dispatcher will look at the events subscribers
and add an event job to each of them.

This adds 2 methods for managing the subscribers to `Client`:
`sync_subscriptions` and `clear_subscriptions`. These methods are
demetered on `Qs` and `Queue` for convenience.

The `sync_subscriptions` method takes a queue and for every event
configured on it, adds the queue name to the events subscribers in
redis. This allows the dispatcher to be aware of the events a
queue is subscribed to and add event jobs to it as needed. This
should be run anytime a queues configured events change.

The `clear_subscriptions` method takes a queue and tries to remove
it from all events subscribers in redis. This ensures that the
queue is no longer subscribed to any events and stops having event
jobs added to it. This should be used when a queue is no longer
being used.