Add optional publisher for events
This adds an optional publisher that can either be configured
globally or specified when publishing an event. This can be used
for tracking purposes and identifying where an event was published
from for debugging. There is now a `publish_as` method which can
be used to specify a publisher when publishing an event. This
method works the same as the `publish` method only it sets the
publisher.