Add `DispatchJobHandler`
This adds a `DispatchJobHandler` mixin which can be used to
define a job handler that handles dispatching events. This handler
is setup to receive `DispatchJob`s and add events to subscribed
queues. This is provided as a mixin so it can be mixed into a
handler where callbacks can be added for custom behavior. This is
part of the events system.

The handler uses `Qs` to read its subscribers from redis and for
each of them it pushes its events payload onto their queue. The
handler is fault tolerant so if an error occurs while pushing an
event onto a queue it will continue trying to push to all queues
before erroring.