Add `Payload` module for serializing and deserializing
This adds a `Payload` module that is responsible for serializing
jobs and deserializing serialized payloads. This doesn't change
any functionality; its just an organization change. It centralizes
all the serialization and deserialization for jobs into one
location and is setup for changing events to inherit from job.

This also renames the Qs serialize and deserialize logic to use
encode and decode. This is to avoid confusion with having multiple
serialize/deserialize methods.