Add `Message`, setup `Event` being a kind of `Message`
This adds a base `Message` class, commonizes logic from the `Job`
class and updates `Job` to inherit from it. This is setup for
`Event` no longer composing a job and instead inheriting from
message.

This is part of a larger effort to remove some of the strange
behavior with events composing jobs. The goal is that both jobs
and events are handled the same when routing them to a handler
but they can be distinct in how they handle their name, params and
other data. We currently have consistent handling with routing but
have awkward logic related to making events and jobs distinct. By
making event a message it can be a distinct class but be routed
consistently as a generic message.

Every kind of message has a payload type, route id and params.
These are the common pieces of data that all messages need to
provide so they can added to queues and routed to handlers. Each
message must provide a route name for its route id.

This also renames many uses of `job` to `message` where the object
may be either an event or job and route name to route id. This is
to be explicit about what the object represents and how it can be
used.