Update `Event` to be a kind of `Message`
This updates `Event` to be a kind of `Message` and makes it a
sibling of `Job`. This is to avoid extra logic working with events
when they composed jobs. This streamlines their handling and
makes them as similar as possible to jobs.

This updates the `Payload` to handle different types of payloads.
It will now use `send` and the payload type to either build a
job/event or their respective payload hashes. To handle unknown
payload types, if this generates an error, it will be rescued
and an invalid payload type error will be raised if the type
isn't one of the two valid types. This allows us to avoid a
conditional checking the type but also try and provide good errors
for most situations.

This also renames usage of job where message is now more
applicable. This was mostly comments but includes a few logs and
variables as well.