Add `MessageHandler`
This adds a `MessageHandler`, moves most of the `JobHandler` logic
into it and then updates `JobHandler` and `EventHandler` to use
it. This is part of making events a kind of message and siblings to
jobs.

The `MessageHandler` contains all the handler-pattern logic for
initializing, the init/run methods and callbacks. It also provides
a params and logger method for all handlers. Both the job and event
handler can use this behavior and the event handler no longer has
to work with the job handlers behavior.

This also extends the job and event handler to provide additional
methods for easily reading their other attributes. These attributes
shouldn't be any harder to access than the params.