Update runners to expect a message
This updates the runners to expect to be passed a generic message.
This is part of making events a kind of message and siblings to
job.

For `Runner` and `QsRunner` this only involved renaming job to
message for the runners and handlers. No other functionality was
changed.

For the test runners, this adds a `TestRunner` for the common
logic between `JobTestRunner` and `EventTestRunner` and changes
`EventTestRunner` to inherit from it. This makes the test runners
siblings which is how jobs and events will eventually be. This
allows them to make use of the common logic for running an handler
in the test suite but also do custom logic as needed.