Pass params via options when building jobs and events
This changes the `new` methods for jobs and events to expect their
params to be passed in an options hash instead of as an argument.
This is to avoid a double-hash at the end of the method signature.

By moving the params into the options, this makes them optional at
the `Job` and `Event` level. Previously, they were optional for
the `enqueue` and `publish` methods. This updates the `Job` and
the `Event` to default them to an empty hash and removes the
previous defaulting.

This also changes the errors thrown by `Job` and `Event` to be
nested under them and renames them to invalid instead of bad. This
is a better naming and lets them use the namespace of their
respective classes for identifying their role.