Add separate logging for jobs and events
This updates the `PayloadHandler` to do custom logging for
different types of messages. Events and jobs will now do different
logging to help differentiate what the message being processed is.
This is to avoid confusion and be explicit about what Qs background
process is doing.

This also changes the `Payload` to have a `type_method_name` method
that converts payload types to method names. This is to validate
that payload types, which can come from the redis DB, do not
execute arbitrary code via `send`. This was previously being done
only for the `Payload` module but the same process is needed for
the `PayloadHandler`. This formalizes the conversion so that the
`PayloadHandler` can make use of it.