Update dispatcher queue
This updates Qs dispatch logic to allow configuring a dispatcher
queue class and job handler class name in addition to a queue name
and job name. This enables Qs to build a working dispatcher queue
that is configured to handle dispatch jobs. It also enables
enabling custom `Queue` behavior by specifying your own queue class
for the dispatcher queue. This is part of the events system.