Benchmark publishing and running events
This updates the benchmark report to also time publishing events
and processing the events. This is so we can be aware of how
changes alter the performance of Qs.

When benchmarking processing events, this is the time it takes for
an event to be dispatched to a queue and then handled by that
queue. This is done the same way that running jobs is benchmarked.
A common IO pipe is written to when a job/event is completed in
a separate process. The report process listens to the IO pipe and
waits for every job/event to complete.