Add daemon event handling system tests
This adds tests to the daemon system tests for processing events.
The tests now spin up a dispatcher daemon and publish events to
the app daemon. This allows testing that a daemon properly handles
events in different scenarios.

This also updates how the keys that are generated by the system
tests are cleaned up. This centralizes all the keys with a `qs-app`
naming so they can easily be found and removed in a teardown block.
This is to avoid a large number of keys building up in a redis
server because the system tests have been run multiple times.