Add subscription system tests, fix sync to remove unneeded subscribers
This adds subscription system tests and fixes a bug in the sync
subscriptions to properly unsubscribe the queue from events if it
no longer listens for them. This is part of the events system and
making sure everything works as expected.

The sync previously just added its queue to its current events
subscribers sets. The issue is if the queue once was listening for
an event but no longer is, it needs to remove itself as a
subscriber of that event. This does that by removing itself from
all events subscribers and then adding itself to the current events
its listening for. This is done in a transaction so there it won't
lose any events while its subscriptions are being synced.