move process label configuration into the daemon
This is more consistent with the process *id* configuration already
being in the daemon.  This also has the benefit of keeping all
non-internal configuration handling being done via the daemon.  The
daemon data object that is passed around now is aware of the process
label.

This option is a little different from the others as it isn't
intended to be driven via a DSL option and also that it can instead
be driven by an ENV var.

The process label attr behaves just as it did before: it still
defaults to the daemon `name`, it can be specified using the same
`QS_PROCESS_LABEL` env var, and it is still applied to the process
name as before.

The goal here is to make the process label value available to the
error handler in a standard way (via the daemon data) just like all
other user-driven configuration values.