hotfix: cleanup block param name on applying worker procs
This should have been done in #93 but was missed.  I had originally
named the methods `*_callbacks` so the `cb` block param made sense.
However, when I renamed the methods to `*_procs` I forgot to update
this block param name accordingly.  I am doing that here now.  This
change doesn't affect the overally behavior - it is just a cleanup
so it can go out with the next version release.

/cc jcredding