Move using the process label env var into the daemon data
This moves using the process label env var into the daemon data.
This is being done because the process label shouldn't be a default
but should be used no matter the configuration. We don't currently
allow configuring a process label, but if we ever do, we want to
make sure the env var is used instead of the configuration.

Functionally, this doesn't change any of qs current behavior. The
process label will use the env var if set and if not it will use
the configured daemon name.

This is related to redding/sanford#162. We did the same changes to
the sanford ip and port env vars. This fixed a bug with the server
not using the ip and port env vars correctly.