message handler: make run callback method public; use in qs runner
This removes some duplication between the message handler and the
qs runner.  The qs runner needs to run some handler callbacks but
was re-implementing the run callback logic.  The handler should
own this logic and be the central location for it.

To fix this, I namespaces the run callback method with `qs_` and
made it public.  Since it is namespaced, I don't feel bad making
it public.  Now that it is public, the qs runner can call it directly.

This is just a cleanup and has no overall behavior changes.  This
is also prep for namespacing all non-user-facing methods in the
message handlers, which I will tackle in a coming effort.