cleanups: consistency with Deas and Sanford; namespace handler meths
This touches up a bunch of minor things to get Qs' handler pattern
implementation/tests consistent with Deas' and Sanford's.

In additon, this namespaces non-user-facing methods in the message
handler.  Since handlers are intended to run user code, we want to be
careful to not "pollute" the handler scope with our own methods that
will break in spectacular ways if the user happens to override them.
This switches the `init` and `run` (ie the two methods not intended
to be called or used by users) to be prefixed with `qs_`.  This not
only marks them as internal "qs" methods but also makes it more
difficult to accidentally override them in custom user handlers.