Update to use the new dat-worker-pool
This updates qs to use the new dat-worker-pool. dat-worker-pool's
interface has changed but functionally it works the same as it
previously did.

A daemon can now be configured with a worker class instead of
worker procs. This is because dat-worker-pool now requires a worker
class when its initialized. A block is no longer part of
dat-worker-pool's interface. The logic for processing a work item
(a queue item for qs) is now part of the worker class that is passed
in. Worker callbacks should now be added to the worker class thats
configured.

The daemon now also builds a worker pool when its initialized
instead of waiting till its started. This was previously made
possible when dat-worker-pool was changed to not start
automatically but qs didn't take advantage of it til now. This
avoids some conditionals and having to set/unset the worker pool
ivar as the daemon is started/stopped.

As part of this change, qs now also provides a `Worker` mixin which
as mentioned before, contains the logic for processing a queue
item. It also contains the on-available and on-error logic that
was previously buried in the `Daemon` class. This mixin is
required for any worker class that is configured on a daemon. The
worker mixin also protects itself from its included hook being run
multiple times to avoid on-error and on-available callbacks being
added multiple times.

This also includes a number of small changes related to using the
new dat-worker-pool:

* Min and max number of workers are no longer configurable because
  dat-worker-pool no longer supports a min or max. A constant
  number of workers is all that can be configured. It now defaults
  to 4 workers (its previous max value).
* Changed the work loop organization to try and make it more
  readable and consistent with dat-tcp/dat-worker-pool.
* Updated styles and conventions to our latest standards
  throughout.