Pass logger to dat-worker-pool when in debug mode
This updates qs debug mode and to pass its logger to
dat-worker-pool when its in debug mode. This allows extra logging
provided by dat-worker-pool. This also removes duplicate logging
that is now handled by dat-worker-pool and the payload handler
raising an exception if in debug mode. All of this is to formalize
the debug mode of qs and take advantage of dat-worker-pools
logging.

The daemon data now has a `debug` flag that is set by the env var
`QS_DEBUG`. When this is set, the qs logger will be passed to
dat-worker-pool. This shouldn't be set in a live environment
because dat-worker-pool does a lot of extra logging that will slow
down qs processing.

This also removes the payload handler raising an exception if the
`QS_DEBUG` flag is set. This is a development feature that was
used when initially developing qs logic. It was never intended to
be a feature outside of the qs repo and test suite. This removes
it so its no longer tied to the `QS_DEBUG` env var. If we need it
again when updating qs we can figure a way to add the feature back
so that it is only part of the test suite.