Rename `Signal` to `State`, use dat-worker-pool locked object
This renames the `Signal` class to `State` which better matches
how the variable is used and avoids confusion with the signal that
we use to wakeup the daemon thread. The `State` class now also
inherits from dat-worker-pool's locked object. This handles the
mutex locking and now the `State` class only implements an
interface for checking what the current state is.