add `queues` method on Daemons
Use this to lookup the queues configured on a daemon.  I want to
use this in some tests to ensure a daemon is listening on the queues
it is supposed to be.

This probably should have been there from the beginning as you can
read other configuration items but was missed.