remove ns-options server config dependency
We are moving away from using ns-options as a dependency b/c it
has performance issues when accessing options and on top of that
its API hasn't proven to be especially useful over a struct-like
class.

This removes ns-options from the daemon config and reworks the
daemon, its config and the daemon data to support the non
ns-options config implementation. The config is now a simple
struct-like class with a bunch of accessors for its attrs.  The
daemon class methods now implement the ns-options-like DSL api
for each of their attrs for backwards compatibility.

Note: I had to rework the system tests and their support files a
bit to get things passing.  This all stems from the daemon system
tests needing to reset daemon configs for each run which isn't
possible now that ns-options is not in play.  I switched to building
a new daemon class on the start of each test run.  This ensures
that the state is the same for each run and also removes the usage
of a global `AppDaemon`.  In addition, this revealed coupling between
tests related to some global support objects.  I kept the global
`AppQueue` as it is used between the queue/daemon system tests but
put it in its own support file.  I switched the support config.qs to
build its own daemon and no longer rely on a global support `AppDaemon`.
Finally, the dispatcher daemon is only needed for the daemon system
tests so I moved its implementation into the tests file.

In addition, I did a number of smaller changes/cleanups, including
some things that aren't backwards compatible:

* switched to calling the config "config" (more succinct than prev
  with no loss of clarity).
* you can no longer set config values to procs (for lazy eval'ing).
  I chose to not keep this behavior.
* I dropped the `to_hash` logic and switched to just manually
  building the daemon data.  I'm just trying to simplify the API
  here and blindly using the config's hash kinda hides what values
  are being set.
* I reworked the order of config/daemon/daemon-data settings to
  try and get everything to match.  This is just to be consistent
  and help keep things organized/readable.  I tried to follow the
  order Deas/Sanford uses as much as possible.
* I did a bunch of test cleanups to not only test the new logic
  but use modern conventions like Factory values and also test
  things more thoroughly.
* updated the assert dependency to get access to the latest stuff.

Overall there are some backwards incompatible changes but the basic
behavior isn't changing.  Most of the diff is just reordering things
and reworking to remove ns-options logic where needed.