remove `dispatcher` namespace from main Qs::Config
This is part of a larger effort to remove ns-optioins as a
dependency of Qs and rework the Config as a general struct object.
We are moving away from using ns-options as a dependency b/c it
has performance issues when accessing options and on top of that
its API hasn't proven to be especially useful over a struct-like
class.

This removes the `dispatcher` namespace and switches its attrs to
be plain accessors on the Config.  I'm just doing this namespace
here to keep the diff from getting too big.  In future efforts
I'll remove the `redis` namespace and remove ns-options from the
Config altogether.

Note: this includes a number of test style cleanups to get up to
our latest conventions.  I noticed these while reworking the
config tests.