remove `redis` namespace from main Qs::Config
This is part of a larger effort to remove ns-options and rework
the Config as a general struct object. We are moving away from
using ns-options as a dependency b/c it has performance issues
when accessing options and on top of that its API hasn't proven
to be especially useful over a struct-like class.

This removes the `redis` namespace and switches its attrs to
be plain accessors on the Config.  I'm just doing this namespace
here to keep the diff from getting too big.  In a future effort
I'll remove ns-options from the Config altogether.

Note there are a couple backwards incompatible changes I want to
call out:

* I renamed the previous `redis_config` method to `redis_connect_hash`.
  This follows the naming in Ardb's `activerecord_connect_hash`
  and is more explicit about that the object is and what its
  purpose is
* I switched to a validate pattern for setting the redis url.  This
  felt more consistent with how the daemon handles its config.
* Since ns-options is no longer in play, I chose to just set the
  config to `nil` on reset.  I also had to switch to manually
  building the redis connect hash.

Note: this also includes removing some commented out code.  This
was accidentally left in from a previous effort (PR 108).