remove ns-options dependency
This is the last part of a larger effort to remove ns-options and
rework the Config as a general struct. We are moving away from
using ns-options as a dependency b/c it has performance issues
when accessing options and on top of that its API hasn't proven
to be especially useful over a struct-like class.

This removes the top-level options and replaces them with generic
accessors.  There are no behavior changes.  This also removes the
ns-options require statements and removes ns-options from the
gemspec.