add `halt` method to message handlers
This method is similar to the `halt` method in Deas view handlers
and Sanford service handlers: it halts execution when called.  A
big difference with Qs' halt method, however, is that it takes no
args and there is no response handling to be concerned with.

This adds the method to the handler (which just proxies the runner
halt method) and implements the method on the runners.  This all
follows the pattern of Sanfords halt method and its tests.