Add `NullClient`, throw uninitialized error when used
This adds a `NullClient` that is the default client for qs if it
hasn't been initialized. This allows us to throw nicer unintialized
errors explaining that qs needs to be init instead of no method
errors.

`Qs` doesn't have a client until its init method is called but it
also demeters its client for convenience. This means that if the
demetered methods are used before `Qs` is init, it would throw
a no method error saying `nil` doesn't respond to the client
method. This changes `Qs` to return a `NullClient` if it hasn't
been initialized. The `NullClient` provides all the methods that
`Qs` demeters but simply raises an uninitialized error whenever
they are called. This lets users know that `Qs` hasn't been
initialized and needs to be before they try enqueueing, publishing,
etc.

Closes #71