Update restart to use ruby 1.9+ `Kernel.exec`
This updates the `Process` restart to use ruby's 1.9+ `Kernel.exec`
interface. This is being done to keep consistent with sanford and
to prep us for supporting future ruby versions.

The `RestartCmd` now has a different run method depending on
which version of ruby is being used. For ruby 1.8.7, its run
method is mostly the same as it was before except it now handles
setting the skip daemonize env var. This also makes the
`RestartCmd` encapsulate all of the logic needed to restart the
daemon. For ruby 1.9+, the run method builds the env variable hash
and builds an exec options hash. The options hash simply tells the
exec to change the dir for now but could be expanded to set other
options if they are needed with future versions of ruby.

This also updates the bench script to work with ruby 1.9+. This
adds the root path to the load path since ruby doesn't do this
automatically anymore.