Hotfix: Update `queues` method to use `config` not `configuration`
This fixes the `queues` method to use `config` instead of
`configuration` and adds unit tests for the method. There weren't
any unit tests so the test suite didn't error and this method isn't
used when running a qs daemon. This was noticed while using qs in
an app because we were testing what queues were configured for a
daemon.